requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community>=0.0.20
google-generativeai>=0.3.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
pypdf==3.17.0
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3

